+++ sparksqlformatter/src/config.py
@@ -22,7 +22,7 @@
 # LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
-from hiveqlformatter.src import hiveql_config as hc
+from sparksqlformatter.src import sparksql_config as sc
 
 
 class Config:
@@ -31,17 +31,10 @@ class Config:
     '''
     def __init__(
             self,
-            keywords=(hc.Keywords.RESERVED_KEYWORDS + hc.Keywords.NON_RESERVED_KEYWORDS +
-                      hc.Functions.MATHEMATICAL_FUNCTIONS + hc.Functions.COLLECTION_FUNCTIONS +
-                      hc.Functions.TYPE_CONVERSION_FUNCTIONS + hc.Functions.DATE_FUNCTIONS +
-                      hc.Functions.CONDITIONAL_FUNCTIONS + hc.Functions.STRING_FUNCTIONS +
-                      hc.Functions.DATA_MASKING_FUNCTIONS + hc.Functions.MISC_FUNCTIONS +
-                      hc.Functions.AGGREGATE_FUNCTIONS + hc.Functions.WINDOWING_FUNCTIONS +
-                      hc.Functions.ANALYTICS_FUNCTIONS),
-            reservedKeywords=hc.Keywords.RESERVED_KEYWORDS,
-            topLevelKeywords=hc.Keywords.TOP_LEVEL_KEYWORDS,
-            topLevelKeywordsNoIndent=hc.Keywords.TOP_LEVEL_KEYWORDS_NO_INDENT,
-            newlineKeywords=hc.Keywords.NEWLINE_KEYWORDS,
+            topLevelKeywords=sc.Keyword.TOP_LEVEL_KEYWORDS,
+            topLevelKeywordsNoIndent=sc.Keyword.TOP_LEVEL_KEYWORDS_NO_INDENT,
+            newlineKeywords=sc.Keyword.NEWLINE_KEYWORDS,
+            userDefinedFunctions=sc.Function.USER_DEFINED_FUNCTIONS,
             stringTypes=['""', "''", '{}'],
             openParens=['(', 'CASE'],
             closeParens=[')', 'END'],  # the order of the parentheses need to match with openParens
@@ -53,10 +46,6 @@ class Config:
             inlineMaxLength=120):
         '''
         Parameters
-        keywords: list
-            Keywords in the query language.
-        reservedKeywords: list
-            Reserved keywords in the query language.
         topLevelKeywords: list
             Keywords that initiate top-level blocks in the query so that the following lines are indented.
             E.g., SELECT and FROM in
@@ -78,6 +67,8 @@ class Config:
                 ...
         newlineKeywords: list
             Keywords that initiate a newline in the query.
+        userDefinedFunctions: list
+            Functions defined by user. To be treated as in-built functions in formatting.
         stringTypes: list
             Pairs of characters that enclose strings in the query.
         openParens: list
@@ -99,8 +90,15 @@ class Config:
         inlineMaxLength: int
             Maximum length of an inline block.
         '''
-        self.keywords = keywords
-        self.reservedKeywords = reservedKeywords
+        self.keywords = (sc.Keyword.RESERVED_KEYWORDS + sc.Keyword.NON_RESERVED_KEYWORDS +
+                         sc.Function.AGGREGATE_FUNCTIONS + sc.Function.ARRAY_FUNCTIONS +
+                         sc.Function.CONDITIONAL_FUNCTIONS + sc.Function.DATE_TIME_FUNCTIONS +
+                         sc.Function.HASH_FUNCTIONS + sc.Function.JSON_FUNCTIONS + sc.Function.MAP_FUNCTIONS +
+                         sc.Function.MATHEMATICAL_FUNCTIONS + sc.Function.MISC_FUNCTIONS +
+                         sc.Function.OPERATOR_FUNCTIONS + sc.Function.STRING_FUNCTIONS + sc.Function.STRUCT_FUNCTIONS +
+                         sc.Function.TABLE_GENERATING_FUNCTIONS + sc.Function.TYPE_CONVERSION_FUNCTIONS +
+                         sc.Function.WINDOW_FUNCTIONS + sc.Function.XPATH_FUNCTIONS + userDefinedFunctions)
+        self.reservedKeywords = sc.Keyword.RESERVED_KEYWORDS
         self.topLevelKeywords = topLevelKeywords
         self.newlineKeywords = newlineKeywords
         self.topLevelKeywordsNoIndent = topLevelKeywordsNoIndent