+++ sparksqlformatter/src/inline_block.py
@@ -22,7 +22,7 @@
 # LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
-from hiveqlformatter.src.tokenizer import TokenType
+from sparksqlformatter.src.tokenizer import TokenType
 
 
 class InlineBlock:
@@ -107,7 +107,7 @@ class InlineBlock:
         Check if token does not belong to an inline block.
 
         Parameters
-        token: hiveqlformatter.src.tokenizer.Token() object
+        token: sparksqlformatter.src.tokenizer.Token() object
 
         Return: bool
         '''