+++ sparksqlformatter/src/subquery.py
@@ -20,7 +20,7 @@
 # LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
-from hiveqlformatter.src.tokenizer import TokenType
+from sparksqlformatter.src.tokenizer import TokenType
 
 
 class SubQuery:
@@ -53,30 +53,38 @@ class SubQuery:
         self.stack = []  # stack to hold unmatched opening and closing parentheses
         self.started = False  # whether a subquery has started, e.g., t0 AS (...)
 
+    def matched(self):
+        '''
+        Check if all parentheses in the subquery have matched.
+
+        Return: bool
+        '''
+        return len(self.stack) == 0
+
     def ended(self):
-        """
+        '''
         Check if the subquery has ended.
 
         Return: bool
-        """
-        return len(self.stack) == 0
+        '''
+        return self.started and len(self.stack) == 0  # subQuery has started and parentheses are matched
 
     def update(self, formatter, token):
-        """
+        '''
         Update self.stack and self.started with current token.
 
         Parameters
-        formatter: hiveqlformatter.src.formatter.Formatter() object
+        formatter: sparksqlformatter.src.formatter.Formatter() object
             The formatter currently in use.
-        token: hiveqlformatter.src.tokenizer.Token() object
+        token: sparksqlformatter.src.tokenizer.Token() object
             The current token.
-        """
+        '''
         openParens = [p for p in formatter.config.openParens if p != 'CASE']
         closeParens = [p for p in formatter.config.closeParens if p != 'END']
         if token.value in openParens:  # push opening parenthesis onto stack
             self.stack.append(token.value)
         elif token.value in closeParens:
-            if self.ended():  # no opening parenthesis to match
+            if self.matched():  # no opening parenthesis to match
                 raise Exception('Parentheses not matched')
             lastParen = self.stack[-1]  # retrieve last opening parenthesis
             if openParens.index(lastParen) == closeParens.index(token.value):  # check if they match
@@ -88,5 +96,4 @@ class SubQuery:
         '''
         Reset the stack and started attributes of the class.
         '''
-        self.stack = []
         self.started = False